Add support for zoom in zoom out buttons defined outside the component
This commit extends the OrgChart component so it's zoom can be controlled by
buttons rendered in the calling code. Developers can use the `controlIds`
property of the OrgChart componet to define html ids of elements that will zoom
in and zoom out the component. The OrgChart component will handle wiring up
onClick events for the provided ids appropriately.